requirements.txt
pytorch==0.4.1
cuda90
tensorboard==1.12.0
numpy==1.15.4
scikit-learn==0.20.2
scipy==1.1.0
cython
pandas==0.24.2
seaborn==0.9.0
tensorflow-gpu==1.12.0
matplotlib==3.0.2
networkx==2.2
jupyterlab
fastprogress==0.1.18
pillow==8.0.1
tensorboardx==1.5
